wuwuwuw

+++ springboot-parent/springboot-service/src/main/java/com/md/service/goods/SpGoodsService.java
@@ -37,7 +37,7 @@ public interface SpGoodsService {
 
     void savePics(String picsUrl) throws Exception;
 
-    Long findGoodsNum() throws Exception;
+    Integer findGoodsNum() throws Exception;
 
     List<Goods> findByName(String query, String pagenum, Integer pagesize) throws Exception;
 }
